requirements.txt
albumentations==0.4.6
matplotlib==3.2.2
PyYAML==5.3.1
seaborn==0.10.1
torch==1.5.1
torchvision==0.6.1
tqdm==4.47.0
